Refactoring Team Event Credit Dashboard into seperate component
