+++ frontend/src/components/Forms/TeamEventCreditsForm/TeamEventCreditsForm.tsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import { Form, Segment, Label, Button, Card, Message } from 'semantic-ui-react';
+import { Form, Segment, Label, Button } from 'semantic-ui-react';
 import { Emitters } from '../../../utils';
 import CustomSearch from '../../Common/Search';
 import { useSelf } from '../../Common/FirestoreDataProvider';
 import { TeamEventsAPI } from '../../../API/TeamEventsAPI';
+import TeamEventCreditDashboard from './TeamEventsCreditsDasboard';
 
 const TeamEventCreditForm: React.FC = () => {
   // When the user is logged in, `useSelf` always return non-null data.
@@ -63,34 +64,6 @@ const TeamEventCreditForm: React.FC = () => {
     }
   };
 
-  const [countApprovedCredits, setCountApprovedCredits] = useState(0);
-  const [countRemainingCredits, setCountRemainingCredits] = useState(3);
-  const [approvedTEC, setApprovedTEC] = useState<TeamEvent[]>([]);
-  const [pendingTEC, setPendingTEC] = useState<TeamEvent[]>([]);
-
-  useEffect(() => {
-    if (teamEvents != null) {
-      teamEvents.forEach((currTeamEvent) => {
-        currTeamEvent.attendees.forEach((currApprovedMember) => {
-          if (currApprovedMember.member.email === userInfo.email) {
-            const currCredits = Number(currTeamEvent.numCredits);
-            setCountApprovedCredits((countApprovedCredits) => countApprovedCredits + currCredits);
-            setCountRemainingCredits((countRemainingCredits) => {
-              if (countRemainingCredits - currCredits <= 0) return 0;
-              return countRemainingCredits - currCredits;
-            });
-            setApprovedTEC((approvedTEC) => [...approvedTEC, currTeamEvent]);
-          }
-        });
-        currTeamEvent.requests.forEach((currApprovedMember) => {
-          if (currApprovedMember.member.email === userInfo.email) {
-            setPendingTEC((pendingTEC) => [...pendingTEC, currTeamEvent]);
-          }
-        });
-      });
-    }
-  }, [teamEvents]);
-
   return (
     <div>
       <Form
@@ -106,7 +79,6 @@ const TeamEventCreditForm: React.FC = () => {
           Earn team event credits for participating in DTI events! Fill out this form every time and
           attach a picture of yourself at the event to receive credit.
         </p>
-
         <div style={{ margin: '2rem 0' }}>
           <label style={{ fontWeight: 'bold' }}>
             Select a Team Event: <span style={{ color: '#db2828' }}>*</span>
@@ -164,7 +136,6 @@ const TeamEventCreditForm: React.FC = () => {
             ) : undefined}
           </div>
         </div>
-
         {teamEvent?.hasHours ? (
           <div>
             <label style={{ fontWeight: 'bold' }}>
@@ -182,7 +153,6 @@ const TeamEventCreditForm: React.FC = () => {
             />
           </div>
         ) : undefined}
-
         <div style={{ margin: '2rem 0' }}>
           <label htmlFor="newImage" style={{ fontWeight: 'bold' }}>
             Upload your event picture here! <span style={{ color: '#db2828' }}>*</span>
@@ -201,59 +171,7 @@ const TeamEventCreditForm: React.FC = () => {
         <Form.Button floated="right" onClick={submitTeamEventCredit}>
           Submit
         </Form.Button>
-
-        <div style={{ margin: '8rem 0' }}></div>
-        <h1>Check Team Event Credits</h1>
-        <p>
-          Check your team event credit status for this semester here! Every DTI member must complete
-          3 team event credits to fulfill this requirement.
-        </p>
-
-        <div style={{ margin: '2rem 0' }}>
-          <label style={{ fontWeight: 'bold' }}>Approved Credits:</label>
-          <p>{countApprovedCredits}</p>
-        </div>
-
-        <div style={{ margin: '2rem 0' }}>
-          <label style={{ fontWeight: 'bold' }}>Approved Events:</label>
-          {approvedTEC.length !== 0 ? (
-            <Card.Group>
-              {approvedTEC.map((teamEvent) => (
-                <Card>
-                  <Card.Content>
-                    <Card.Header>{teamEvent.name} </Card.Header>
-                    <Card.Meta>{teamEvent.date}</Card.Meta>
-                  </Card.Content>
-                </Card>
-              ))}
-            </Card.Group>
-          ) : (
-            <Message>You have not been approved for any team events yet.</Message>
-          )}
-        </div>
-
-        <div style={{ margin: '2rem 0' }}>
-          <label style={{ fontWeight: 'bold' }}>Remaining Credits Needed:</label>
-          <p>{countRemainingCredits}</p>
-        </div>
-
-        <div style={{ margin: '2rem 0' }}>
-          <label style={{ fontWeight: 'bold' }}>Pending Approval For::</label>
-          {pendingTEC.length !== 0 ? (
-            <Card.Group>
-              {pendingTEC.map((teamEvent) => (
-                <Card>
-                  <Card.Content>
-                    <Card.Header>{teamEvent.name} </Card.Header>
-                    <Card.Meta>{teamEvent.date}</Card.Meta>
-                  </Card.Content>
-                </Card>
-              ))}
-            </Card.Group>
-          ) : (
-            <Message>You are not currently pending approval for any team events.</Message>
-          )}
-        </div>
+        <TeamEventCreditDashboard userInfo={userInfo} />
       </Form>
     </div>
   );
